Trainer can view bookings, reviews and who booked each training
